cronjob/cron5/requirements.txt
requests==2.26.0
mysql-connector-python==8.0.27
pyyml
cython==0.29.24